requirements.txt
django-bootstrap4-datetimepicker==4.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.1.1
django-crispy-forms==1.9.2

